requirements/production.txt
# Production dependencies
-r base.txt

# Environment management
django-environ>=0.11,<1.0

# PostgreSQL adapter
psycopg2-binary>=2.9,<3.0

# Production WSGI server
gunicorn>=22.0,<23.0

# Static file serving
whitenoise>=6.7,<7.0

# Redis client
redis>=5.0,<6.0

# Celery for background tasks
celery>=5.3,<6.0
django-celery-beat>=2.5,<3.0
django-celery-results>=2.5,<3.0
